Added code for image normalization over entire training data
This patch normalizes each image feature to get a z-score.
Without this, if each feature has a different range
of intensities, the weight vectors of NN have to be optimized
for each of these different range of distributions. As we only use
a single global learning rate to update these vectors, it is ideal
to have all features in the same range.